Opção de remover aparece apenas quando não há um próximo ponto

+++ lib/features/widgets/dialog_edit_point.dart
@@ -39,17 +39,18 @@ Future dialogEditPoint(
               style: TextStyle(color: Colors.redAccent),
             ),
           ),
-          TextButton(
-            onPressed: () {
-              points.remove(point);
-              PointRepository().deletePoint(token, point.uuid);
-              Navigator.pop(context);
-            },
-            child: const Text(
-              "Remover",
-              style: TextStyle(color: Colors.redAccent),
+          if (point.neighbor["next"] == null)
+            TextButton(
+              onPressed: () {
+                points.remove(point);
+                PointRepository().deletePoint(token, point.uuid);
+                Navigator.pop(context);
+              },
+              child: const Text(
+                "Remover",
+                style: TextStyle(color: Colors.redAccent),
+              ),
             ),
-          ),
           TextButton(
             onPressed: () {
               dialogEditar(context, point);
